mini-x-session: add .bbappend file
If use x11-base IMAGE_FEATURES when build the image, the mini-x-session
will launch a terminal window on startup, which will hide jivelite
window.

Tweak the mini-x-session script to not start the terminal and allow
jivelite to run fullscreen without cursor displayed, this has better
user experience on touch screen. :)

Signed-off-by: Zhu Yi <dev.yi.zhu@gmail.com>